+++ CRM/CivirulesConditions/Form/ContributionRecur/DonorIsRecurring.php
@@ -1,12 +1,23 @@
 <?php
+
+use CRM_Civirules_ExtensionUtil as E;
 /**
  * Class for CiviRules Condition Contribution Donor Is Recurring
  *
  * @author Erik Hommel (CiviCooP) <erik.hommel@civicoop.org>
  * @license AGPL-3.0
  */
-
 class CRM_CivirulesConditions_Form_ContributionRecur_DonorIsRecurring extends CRM_CivirulesConditions_Form_Form {
+
+  /**
+   * Returns a help text for this condition.
+   * The help text is shown to the administrator who is configuring the condition.
+   *
+   * @return string
+   */
+  protected function getHelpText() {
+    return E::ts('This condition checks if the contact has any recurring contributions with no end-date or and end-date later than today. It does NOT check the status of a recurring contribution and does not work with test entities.');
+  }
 
   /**
    * Overridden parent method to build form
@@ -25,7 +36,6 @@ class CRM_CivirulesConditions_Form_ContributionRecur_DonorIsRecurring extends CR
    * Overridden parent method to set default values
    *
    * @return array $defaultValues
-   * @access public
    */
   public function setDefaultValues() {
     $defaultValues = parent::setDefaultValues();
@@ -40,7 +50,6 @@ class CRM_CivirulesConditions_Form_ContributionRecur_DonorIsRecurring extends CR
    * Overridden parent method to process form data after submission
    *
    * @throws Exception when rule condition not found
-   * @access public
    */
   public function postProcess() {
     if (isset($this->_submitValues['has_recurring'])) {
@@ -53,4 +62,5 @@ class CRM_CivirulesConditions_Form_ContributionRecur_DonorIsRecurring extends CR
 
     parent::postProcess();
   }
+
 }